// src/payload-types.ts
export interface Guest {
  id: number;
  name: string;
  done: 'yes' | 'no';
  attending?: ('yes' | 'no') | null;
  maxGuests: number;
  numOfRsvp?: number | null;
  dietaryRestrictions?: string | null;
  slugParam?: string | null;
  updatedAt: string;
  createdAt: string;
}